backend/requirements.txt
fastapi
uvicorn
pandas
matplotlib
seaborn
Jinja2
httpx
python-multipart
sqlalchemy
python-jose[cryptography]
passlib
python-dotenv
bcrypt
plotly
scikit-learn
statsmodels
psycopg2-binary
pymysql
boto3
pdfplumber
openpyxl
email-validator
mysql-connector-python